Split logic into functions and separate into modules
